Replace long by int in qi::atomic
Reviewed-on: http://gerrit.aldebaran.lan:8080/8417
Tested-by: jfreche <jfreche@aldebaran-robotics.com>

+++ src/message.cpp
@@ -22,7 +22,7 @@ namespace qi {
 
   unsigned int newMessageId()
   {
-    static qi::Atomic<long> id(0);
+    static qi::Atomic<int> id(0);
     return ++id;
   }
